create the db schema
