+++ package.json
@@ -7,6 +7,7 @@
     "@eslint/js": "^10.0.1",
     "@stylistic/eslint-plugin": "^5.7.1",
     "@types/bun": "latest",
+    "drizzle-kit": "^1.0.0-beta.15-859cf75",
     "eslint": "^10.0.0",
     "globals": "^17.3.0",
     "jiti": "^2.6.1",
@@ -18,6 +19,8 @@
   "dependencies": {
     "@elysiajs/bearer": "^1.4.3",
     "@elysiajs/eden": "^1.4.8",
-    "elysia": "^1.4.22"
+    "drizzle-orm": "^1.0.0-beta.15-859cf75",
+    "elysia": "^1.4.22",
+    "nanoid": "^5.1.6"
   }
 }
